Moved ParamCurveBezierClipping into new file
